backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-cors-headers==3.13.0
django-enum==1.1.1
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
nordigen==1.3.0
packaging==23.0
pluggy==1.0.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
